boss-rally/BRD3D: update decomp

+++ boss-rally/BRD3D/split-win32/sub_10060/sub_10060260.c
@@ -1,5 +1,5 @@
 #include "types-win32.h"
 //----- (10060260) --------------------------------------------------------
 LPDIRECTINPUTDEVICEA sub_10060260() {
-  return meth_unk00054_100603A0(gUnk54Ptr, (int)gHwnd);
+  return meth_unk00054_100603A0(gUnk54Ptr, gHwnd);
 }
